Added input variable for Strings
We can input our own strings from the serialtest4.py script now which will be sent to Arduino.
Arduino now has a byte array in which it collects all the incoming bytes and prints it to the serial port.

There's probably a max of 512 bytes of strings to send so keep that in mind!